third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/media-elements/error-sequence-expected.txt

This is a testharness.js-based test.
[FAIL] error event while HAVE_NOTHING
  assert_equals: Expected error event, but got volumechange event instead expected "error" but got "volumechange"
[FAIL] play() promise while HAVE_NOTHING
  assert_array_equals: sequence expected property 1 to be "rejection" but got "volumechange" (expected array ["volumechange", "rejection", "volumechange"] got ["volumechange", "volumechange", "rejection"])
[FAIL] error event while HAVE_METADATA
  assert_equals: Expected volumechange event, but got pause event instead expected "volumechange" but got "pause"
[FAIL] error attribute after DECODE_ERROR
  assert_equals: error code expected (number) 3 but got (undefined) undefined
Harness: the test ran to completion.
